Refactor `http_api_helpers.py` and `response.py`
